+++ doc/whats_new.rst
@@ -9,6 +9,16 @@ New features
     - plotting plot_connectome with colorbar is now implemented in function
       :func:`nilearn.plotting.plot_connectome`
 
+API changes summary
+...................
+    - Atlas fetcher :func:`nilearn.datasets.fetch_atlas_msdl` now returns directly
+      labels of the regions in output variable 'labels' and its coordinates
+      in output variable 'region_coords' and its type of network in 'networks'.
+    - The output variable name 'regions' is now changed to 'maps' in AAL atlas
+      fetcher in :func:`nilearn.datasets.fetch_atlas_aal`.
+    - AAL atlas now returns directly its labels in variable 'labels' and its
+      index values in variable 'indices'.
+
 0.2.3
 =====